fix: cross compilation API
Closes https://github.com/apache/camel-k/issues/4531
