+++ pkg/builder/build_test.go
@@ -4,9 +4,7 @@ import (
 	"archive/tar"
 	"fmt"
 	"os"
-	"regexp"
 	"strings"
-	"syscall"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -73,32 +71,6 @@ func TestTarSingleEntry(t *testing.T) {
 	assert.NotEqual(t, 0, header.Gid)
 }
 
-func TestTarPermissionsSingleEntry(t *testing.T) {
-	var tmpFile1 *os.File
-	var err error
-	if tmpFile1, err = os.CreateTemp("", "camel-k-*.txt"); err != nil {
-		t.Error(err)
-	}
-
-	assert.Nil(t, tmpFile1.Close())
-	assert.Nil(t, os.WriteFile(tmpFile1.Name(), []byte(`
-	This is for simple testing
-	`), 0o400))
-
-	tarFileName, err := tarPackage(tmpFile1.Name(), "/path/to/target", false)
-	assert.NoError(t, err)
-	r, err := os.Open(tarFileName)
-	assert.NoError(t, err)
-	tr := tar.NewReader(r)
-	assert.NotNil(t, tr)
-	header, err := tr.Next()
-	assert.NotNil(t, tr)
-	assert.True(t, strings.HasPrefix(header.Name, "/path/to/target/camel-k-"))
-	assert.True(t, strings.HasSuffix(header.Name, ".txt"))
-	assert.Equal(t, syscall.Getuid(), header.Uid)
-	assert.Equal(t, syscall.Getgid(), header.Gid)
-}
-
 func TestTarDir(t *testing.T) {
 	var tmpDir string
 	var tmpFile1 *os.File
@@ -157,64 +129,4 @@ func TestTarDirRecursive(t *testing.T) {
 	assert.NotNil(t, tr)
 	assert.True(t, strings.HasPrefix(header.Name, "/path/to/target/dir2"))
 	assert.True(t, strings.HasSuffix(header.Name, ".txt"))
-}
-
-func TestTarDirAllEntriesRecursive(t *testing.T) {
-	tmpDir1, err := os.MkdirTemp("", "camel-k-dir1-*")
-	assert.NoError(t, err)
-	tmpDir2, err := os.MkdirTemp(tmpDir1, "camel-k-dir2-*")
-	assert.NoError(t, err)
-	_, err = os.MkdirTemp(tmpDir1, "camel-k-dir3-*")
-	assert.NoError(t, err)
-	tmpFile1, err := os.CreateTemp(tmpDir2, "camel-k-file-*.txt")
-	assert.NoError(t, err)
-
-	assert.Nil(t, tmpFile1.Close())
-	assert.Nil(t, os.WriteFile(tmpFile1.Name(), []byte(`
-	This is for simple testing
-	`), 0o400))
-
-	tarFileName, err := tarPackage(tmpDir1, "/path/to/target", true)
-	assert.NoError(t, err)
-	r, err := os.Open(tarFileName)
-	assert.NoError(t, err)
-	tr := tar.NewReader(r)
-	assert.NotNil(t, tr)
-
-	header, err := tr.Next()
-	assert.NotNil(t, header)
-	assert.Nil(t, err)
-	matched, _ := regexp.MatchString("/path/to/target/", header.Name)
-	assert.True(t, matched)
-	assert.Equal(t, syscall.Getuid(), header.Uid)
-	assert.Equal(t, syscall.Getgid(), header.Gid)
-
-	header, err = tr.Next()
-	assert.NotNil(t, header)
-	assert.Nil(t, err)
-	matched, _ = regexp.MatchString("/path/to/target/camel-k-dir2-.*", header.Name)
-	assert.True(t, matched)
-	assert.Equal(t, syscall.Getuid(), header.Uid)
-	assert.Equal(t, syscall.Getgid(), header.Gid)
-
-	header, err = tr.Next()
-	assert.NotNil(t, header)
-	assert.Nil(t, err)
-	matched, _ = regexp.MatchString("/path/to/target/camel-k-dir2-.*/camel-k-file-.*\\.txt", header.Name)
-	assert.True(t, matched)
-	assert.Equal(t, syscall.Getuid(), header.Uid)
-	assert.Equal(t, syscall.Getgid(), header.Gid)
-
-	header, err = tr.Next()
-	assert.NotNil(t, header)
-	assert.Nil(t, err)
-	matched, _ = regexp.MatchString("/path/to/target/camel-k-dir3-.*/", header.Name)
-	assert.True(t, matched)
-	assert.Equal(t, syscall.Getuid(), header.Uid)
-	assert.Equal(t, syscall.Getgid(), header.Gid)
-
-	header, err = tr.Next()
-	assert.Nil(t, header)
-	assert.NotNil(t, err)
-	assert.Equal(t, "EOF", err.Error())
 }
